+++ apps/web/tsconfig.json
@@ -9,6 +9,5 @@
     "lib": ["ESNext", "DOM", "DOM.Iterable"],
     "types": ["react", "react-dom", "vite/client"]
   },
-  "include": ["src"],
-  "exclude": ["**/*.test.ts", "**/*.test.tsx"]
+  "include": ["src"]
 }
