+++ packages/contracts/src/rpc.ts
@@ -1,1 +1,5 @@
 export const RPC_PREFIX = "/rpc";
+
+export function joinRpcUrl(baseUrl: string): string {
+  return `${baseUrl.replace(/\/$/, "")}${RPC_PREFIX}`;
+}
